chore: fix clippy warnings and improve code quality
- Replace manual prefix stripping with strip_prefix() in TUI commands
- Simplify identical if blocks for TOOL: vs TOOL. parsing
- Use .copied() instead of .map(|s| *s) for char iteration
- Remove redundant String::from(format!()) conversions
- Replace unwrap_or_else with const fallbacks where possible
- Remove unnecessary closures in unwrap_or_else calls
- Use div_ceil() for ceiling division instead of manual arithmetic
- Clean up duplicate #[allow(dead_code)] attributes
- Fix empty lines after doc comments
- Remove useless format!() around static strings
- Use inline format captures in println!() args
- Collapse redundant if/else if branches with identical returns
- Replace sort_by with sort_by_key for cleaner comparisons
- Use !is_empty() instead of .len() > 0
- Remove redundant .trim() before .split_whitespace()
- Remove unnecessary as u64 casts
- Fix let _ = on Result-returning calls
- Derive Default instead of manual impl
- Use enumerate().take() instead of index loops
- Remove Ok(...) wrappers around ? expressions
- Use .values() instead of iter().map() for map iteration

Fixes 48 clippy warnings (77 -> 29 remaining)
All 363 tests pass, build stays green

+++ src/gateway/telegram.rs
@@ -129,7 +129,7 @@ async fn handle_message(
         })
         .unwrap_or_else(|| "Unknown".to_string());
 
-    let user_id = user.map(|u| u.id.0 as u64).unwrap_or(0);
+    let user_id = user.map(|u| u.id.0).unwrap_or(0);
 
     let event = TelegramEvent::UserMessage {
         chat_id,
